+++ packages/react-dom/src/__tests__/ReactDOMServerIntegrationRefs-test.js
@@ -1,0 +1,108 @@
+/**
+ * Copyright (c) 2013-present, Facebook, Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @emails react-core
+ */
+
+'use strict';
+
+const ReactDOMServerIntegrationUtils = require('./utils/ReactDOMServerIntegrationTestUtils');
+
+let React;
+let ReactDOM;
+let ReactDOMServer;
+
+// When there is a test that renders on server and then on client and expects a logged
+// error, you want to see the error show up both on server and client. Unfortunately,
+// React refuses to issue the same error twice to avoid clogging up the console.
+// To get around this, we must reload React modules in between server and client render.
+function initModules() {
+  // First, reset the modules to load the client renderer.
+  jest.resetModuleRegistry();
+  React = require('react');
+  ReactDOM = require('react-dom');
+
+  // Now we reset the modules again to load the server renderer.
+  // Resetting is important because we want to avoid any shared state
+  // influencing the tests.
+  jest.resetModuleRegistry();
+  ReactDOMServer = require('react-dom/server');
+
+  // Make them available to the helpers.
+  return {
+    ReactDOM,
+    ReactDOMServer,
+  };
+}
+
+const {
+  resetModules,
+  asyncReactDOMRender,
+  clientRenderOnServerString,
+  expectMarkupMatch,
+} = ReactDOMServerIntegrationUtils(initModules);
+
+describe('ReactDOMServerIntegration', () => {
+  beforeEach(() => {
+    resetModules();
+  });
+
+  describe('refs', function() {
+    it('should not run ref code on server', async () => {
+      let refCount = 0;
+      class RefsComponent extends React.Component {
+        render() {
+          return <div ref={e => refCount++} />;
+        }
+      }
+      await expectMarkupMatch(<RefsComponent />, <div />);
+      expect(refCount).toBe(0);
+    });
+
+    it('should run ref code on client', async () => {
+      let refCount = 0;
+      class RefsComponent extends React.Component {
+        render() {
+          return <div ref={e => refCount++} />;
+        }
+      }
+      await expectMarkupMatch(<div />, <RefsComponent />);
+      expect(refCount).toBe(1);
+    });
+
+    it('should send the correct element to ref functions on client', async () => {
+      let refElement = null;
+      class RefsComponent extends React.Component {
+        render() {
+          return <div ref={e => (refElement = e)} />;
+        }
+      }
+      const e = await clientRenderOnServerString(<RefsComponent />);
+      expect(refElement).not.toBe(null);
+      expect(refElement).toBe(e);
+    });
+
+    it('should have string refs on client when rendered over server markup', async () => {
+      class RefsComponent extends React.Component {
+        render() {
+          return <div ref="myDiv" />;
+        }
+      }
+
+      const markup = ReactDOMServer.renderToString(<RefsComponent />);
+      const root = document.createElement('div');
+      root.innerHTML = markup;
+      let component = null;
+      resetModules();
+      await asyncReactDOMRender(
+        <RefsComponent ref={e => (component = e)} />,
+        root,
+        true,
+      );
+      expect(component.refs.myDiv).toBe(root.firstChild);
+    });
+  });
+});
